Add basic mesh classes: MeshVertex and MeshFace.
Initial mesh support + refactoring (float -> double types).
MeshVertex class:
  1. 3D coordinates
  2. Normal vector
  3. Set of corresponding faces (by indexes).

MeshFace class (triangle):
  1. 3 corresponding vertexes (by indexes).
  2. Ray intersection handling
  3. Interpolation of normal vectors.

Mesh class (initial):
  1. Array of indexes
  2. Array of faces

+++ lib/Camera.cpp
@@ -1,6 +1,6 @@
 #include "Camera.h"
 
-Camera::Camera(const glm::vec3 &pos, const glm::vec3 &dir,
+Camera::Camera(const glm::dvec3 &pos, const glm::dvec3 &dir,
 							 const glm::vec2 res)
 	: position(pos), direction(dir), resolution(res) {
 	Normalize();
@@ -8,15 +8,15 @@ Camera::Camera(const glm::vec3 &pos, const glm::vec3 &dir,
 
 Camera::~Camera() {}
 
-void Camera::LookAt(const glm::vec3 &point) {
+void Camera::LookAt(const glm::dvec3 &point) {
 	if (point != position) {
 		direction = point - position;
 		Normalize();
 	}
 }
 
-void Camera::MoveTo(const glm::vec3 &point) {
-	glm::vec3 focusPoint = position + direction;
+void Camera::MoveTo(const glm::dvec3 &point) {
+	glm::dvec3 focusPoint = position + direction;
 	position = point;
 	if (point != focusPoint) {
 		// If we try to move to the focus point, keep the direction.
@@ -26,20 +26,20 @@ void Camera::MoveTo(const glm::vec3 &point) {
 	Normalize();
 }
 
-void Camera::MoveForward(float distance) {
-	assert(std::abs(glm::length(direction) - 1.0f) < 0.00001f &&
+void Camera::MoveForward(double distance) {
+	assert(std::abs(glm::length(direction) - 1.0) < 0.00001 &&
 				 "Camera normalization failed!");
 	position += distance * direction;
 }
 
-void Camera::MoveBackward(float distance) {
-	assert(std::abs(glm::length(direction) - 1.0f) < 0.00001f &&
+void Camera::MoveBackward(double distance) {
+	assert(std::abs(glm::length(direction) - 1.0) < 0.00001 &&
 				 "Camera normalization failed!");
 	position -= distance * direction;
 }
 
 void Camera::Normalize() {
 	direction = glm::normalize(direction);
-	assert(std::abs(glm::length(direction) - 1.0f) < 0.00001f &&
+	assert(std::abs(glm::length(direction) - 1.0) < 0.00001 &&
 				 "Camera normalization failed!");
 }
