+++ lib/Camera.h
@@ -7,37 +7,37 @@ class Camera {
 public:
 	// === Constructors ===
 
-	Camera(const glm::vec3 &pos, const glm::vec3 &dir, const glm::vec2 res);
+	Camera(const glm::dvec3 &pos, const glm::dvec3 &dir, const glm::vec2 res);
 	~Camera();
 
 	// === Camera movement ===
 
 	// Change camera's focus to \p point, preserving the position.
 	// Normalizes direction.
-	void LookAt(const glm::vec3 &point);
+	void LookAt(const glm::dvec3 &point);
 
 	// Change camera's position to \p point preserving the focus.
 	// Normalizes direction.
-	void MoveTo(const glm::vec3 &point);
+	void MoveTo(const glm::dvec3 &point);
 
 	// Move camera forward (collinear to direction).
-	void MoveForward(float distance);
+	void MoveForward(double distance);
 
 	// Move camera backward (opposite to direction).
-	void MoveBackward(float distance);
+	void MoveBackward(double distance);
 
 private:
 	// Normalize direction - make it of length 1.
 	void Normalize();
 
-	glm::vec3 position;
-	glm::vec3 direction;
+	glm::dvec3 position;
+	glm::dvec3 direction;
 	glm::uvec2 resolution;
 
 public:
 	// Getters.
-	glm::vec3 getPosition() const { return position; }
-	glm::vec3 getDirection() const { return direction; }
+	glm::dvec3 getPosition() const { return position; }
+	glm::dvec3 getDirection() const { return direction; }
 	glm::uvec2 getResolution() const { return resolution; }
 };
